+++ server/bundled/tools/python/ast_parser.py
@@ -111,6 +111,8 @@ class Analyzer(ast.NodeVisitor):
             'end_col_offset': kwargs.get('end_col_offset'),
             'is_reserved': kwargs.get('is_reserved', False),
             'has_set_foreign_key_related_name': kwargs.get('has_set_foreign_key_related_name', None),
+            'has_set_foreign_key_on_delete': kwargs.get('has_set_foreign_key_on_delete', None),
+            "is_charfield_or_textfield_nullable": kwargs.get("is_charfield_or_textfield_nullable", None),
         }
         
         if 'value' in kwargs:
@@ -135,6 +137,8 @@ class Analyzer(ast.NodeVisitor):
             symbol['arguments'] = kwargs['arguments']
         if kwargs.get('high_priority'):
             symbol['high_priority'] = kwargs['high_priority']
+        if kwargs.get('target_positions'):
+            symbol['target_positions'] = kwargs['target_positions']
         if kwargs.get('full_line_length'):
             symbol['full_line_length'] = kwargs['full_line_length']
         
@@ -399,6 +403,7 @@ class Analyzer(ast.NodeVisitor):
         self.generic_visit(node)
 
     def visit_For(self, node):
+        LOGGER.error(f"Visiting for loop {node}")
         comments = self.get_related_comments(node)
         target = None
         target_positions = []